elt_snowflake/requirements.txt
cx_Oracle
python-decouple
python-dotenv
pandas
snowflake-connector-python
